+++ float64validator/at_most.go
@@ -7,29 +7,28 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/hashicorp/terraform-plugin-framework/function"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 
 	"github.com/hashicorp/terraform-plugin-framework-validators/helpers/validatordiag"
+	"github.com/hashicorp/terraform-plugin-framework-validators/helpers/validatorfuncerr"
 )
 
 var _ validator.Float64 = atMostValidator{}
+var _ function.Float64ParameterValidator = atMostValidator{}
 
-// atMostValidator validates that an float Attribute's value is at most a certain value.
 type atMostValidator struct {
 	max float64
 }
 
-// Description describes the validation in plain text formatting.
 func (validator atMostValidator) Description(_ context.Context) string {
 	return fmt.Sprintf("value must be at most %f", validator.max)
 }
 
-// MarkdownDescription describes the validation in Markdown formatting.
 func (validator atMostValidator) MarkdownDescription(ctx context.Context) string {
 	return validator.Description(ctx)
 }
 
-// ValidateFloat64 performs the validation.
 func (v atMostValidator) ValidateFloat64(ctx context.Context, request validator.Float64Request, response *validator.Float64Response) {
 	if request.ConfigValue.IsNull() || request.ConfigValue.IsUnknown() {
 		return
@@ -46,15 +45,31 @@ func (v atMostValidator) ValidateFloat64(ctx context.Context, request validator.
 	}
 }
 
+func (v atMostValidator) ValidateParameterFloat64(ctx context.Context, request function.Float64ParameterValidatorRequest, response *function.Float64ParameterValidatorResponse) {
+	if request.Value.IsNull() || request.Value.IsUnknown() {
+		return
+	}
+
+	value := request.Value.ValueFloat64()
+
+	if value > v.max {
+		response.Error = validatorfuncerr.InvalidParameterValueFuncError(
+			request.ArgumentPosition,
+			v.Description(ctx),
+			fmt.Sprintf("%f", value),
+		)
+	}
+}
+
 // AtMost returns an AttributeValidator which ensures that any configured
-// attribute value:
+// attribute or function parameter value:
 //
 //   - Is a number, which can be represented by a 64-bit floating point.
 //   - Is less than or equal to the given maximum.
 //
 // Null (unconfigured) and unknown (known after apply) values are skipped.
-func AtMost(max float64) validator.Float64 {
+func AtMost(maxVal float64) atMostValidator {
 	return atMostValidator{
-		max: max,
+		max: maxVal,
 	}
 }