+++ float64validator/none_of_test.go
@@ -5,8 +5,10 @@ package float64validator_test
 
 import (
 	"context"
+	"fmt"
 	"testing"
 
+	"github.com/hashicorp/terraform-plugin-framework/function"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 
@@ -17,71 +19,82 @@ func TestNoneOfValidator(t *testing.T) {
 	t.Parallel()
 
 	type testCase struct {
-		in        types.Float64
-		validator validator.Float64
-		expErrors int
+		in           types.Float64
+		noneOfValues []float64
+		expectError  bool
 	}
 
 	testCases := map[string]testCase{
 		"simple-match": {
 			in: types.Float64Value(123.456),
-			validator: float64validator.NoneOf(
+			noneOfValues: []float64{
 				123.456,
 				234.567,
 				8910.11,
 				1213.1415,
-			),
-			expErrors: 1,
+			},
+			expectError: true,
 		},
 		"simple-mismatch": {
 			in: types.Float64Value(123.456),
-			validator: float64validator.NoneOf(
+			noneOfValues: []float64{
 				234.567,
 				8910.11,
 				1213.1415,
-			),
-			expErrors: 0,
+			},
 		},
 		"skip-validation-on-null": {
 			in: types.Float64Null(),
-			validator: float64validator.NoneOf(
+			noneOfValues: []float64{
 				234.567,
 				8910.11,
 				1213.1415,
-			),
-			expErrors: 0,
+			},
 		},
 		"skip-validation-on-unknown": {
 			in: types.Float64Unknown(),
-			validator: float64validator.NoneOf(
+			noneOfValues: []float64{
 				234.567,
 				8910.11,
 				1213.1415,
-			),
-			expErrors: 0,
+			},
 		},
 	}
 
 	for name, test := range testCases {
 		name, test := name, test
-		t.Run(name, func(t *testing.T) {
+
+		t.Run(fmt.Sprintf("ValidateFloat64 - %s", name), func(t *testing.T) {
 			t.Parallel()
 			req := validator.Float64Request{
 				ConfigValue: test.in,
 			}
 			res := validator.Float64Response{}
-			test.validator.ValidateFloat64(context.TODO(), req, &res)
+			float64validator.NoneOf(test.noneOfValues...).ValidateFloat64(context.TODO(), req, &res)
 
-			if test.expErrors > 0 && !res.Diagnostics.HasError() {
-				t.Fatalf("expected %d error(s), got none", test.expErrors)
+			if !res.Diagnostics.HasError() && test.expectError {
+				t.Fatal("expected error, got no error")
 			}
 
-			if test.expErrors > 0 && test.expErrors != res.Diagnostics.ErrorsCount() {
-				t.Fatalf("expected %d error(s), got %d: %v", test.expErrors, res.Diagnostics.ErrorsCount(), res.Diagnostics)
+			if res.Diagnostics.HasError() && !test.expectError {
+				t.Fatalf("got unexpected error: %s", res.Diagnostics)
+			}
+		})
+
+		t.Run(fmt.Sprintf("ValidateParameterFloat64 - %s", name), func(t *testing.T) {
+			t.Parallel()
+			req := function.Float64ParameterValidatorRequest{
+				Value: test.in,
+			}
+			res := function.Float64ParameterValidatorResponse{}
+			float64validator.NoneOf(test.noneOfValues...).ValidateParameterFloat64(context.TODO(), req, &res)
+
+			if res.Error == nil && test.expectError {
+				t.Fatal("expected error, got no error")
 			}
 
-			if test.expErrors == 0 && res.Diagnostics.HasError() {
-				t.Fatalf("expected no error(s), got %d: %v", res.Diagnostics.ErrorsCount(), res.Diagnostics)
+			if res.Error != nil && !test.expectError {
+				t.Fatalf("got unexpected error: %s", res.Error)
 			}
 		})
 	}