+++ int32validator/between_test.go
@@ -5,8 +5,10 @@ package int32validator_test
 
 import (
 	"context"
+	"fmt"
 	"testing"
 
+	"github.com/hashicorp/terraform-plugin-framework/function"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
@@ -61,11 +63,18 @@ func TestBetweenValidator(t *testing.T) {
 			max:         3,
 			expectError: true,
 		},
+		"invalid validator usage - minVal > maxVal": {
+			val:         types.Int32Value(2),
+			min:         3,
+			max:         1,
+			expectError: true,
+		},
 	}
 
 	for name, test := range tests {
 		name, test := name, test
-		t.Run(name, func(t *testing.T) {
+
+		t.Run(fmt.Sprintf("ValidateInt32 - %s", name), func(t *testing.T) {
 			t.Parallel()
 			request := validator.Int32Request{
 				Path:           path.Root("test"),
@@ -81,6 +90,24 @@ func TestBetweenValidator(t *testing.T) {
 
 			if response.Diagnostics.HasError() && !test.expectError {
 				t.Fatalf("got unexpected error: %s", response.Diagnostics)
+			}
+		})
+
+		t.Run(fmt.Sprintf("ValidateParameterInt32 - %s", name), func(t *testing.T) {
+			t.Parallel()
+			request := function.Int32ParameterValidatorRequest{
+				ArgumentPosition: 0,
+				Value:            test.val,
+			}
+			response := function.Int32ParameterValidatorResponse{}
+			int32validator.Between(test.min, test.max).ValidateParameterInt32(context.TODO(), request, &response)
+
+			if response.Error == nil && test.expectError {
+				t.Fatal("expected error, got no error")
+			}
+
+			if response.Error != nil && !test.expectError {
+				t.Fatalf("got unexpected error: %s", response.Error)
 			}
 		})
 	}