+++ int64validator/at_most_test.go
@@ -5,8 +5,10 @@ package int64validator_test
 
 import (
 	"context"
+	"fmt"
 	"testing"
 
+	"github.com/hashicorp/terraform-plugin-framework/function"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
@@ -48,7 +50,8 @@ func TestAtMostValidator(t *testing.T) {
 
 	for name, test := range tests {
 		name, test := name, test
-		t.Run(name, func(t *testing.T) {
+
+		t.Run(fmt.Sprintf("ValidateInt64 - %s", name), func(t *testing.T) {
 			t.Parallel()
 			request := validator.Int64Request{
 				Path:           path.Root("test"),
@@ -64,6 +67,23 @@ func TestAtMostValidator(t *testing.T) {
 
 			if response.Diagnostics.HasError() && !test.expectError {
 				t.Fatalf("got unexpected error: %s", response.Diagnostics)
+			}
+		})
+
+		t.Run(fmt.Sprintf("ValidateParameterInt64 - %s", name), func(t *testing.T) {
+			t.Parallel()
+			request := function.Int64ParameterValidatorRequest{
+				Value: test.val,
+			}
+			response := function.Int64ParameterValidatorResponse{}
+			int64validator.AtMost(test.max).ValidateParameterInt64(context.TODO(), request, &response)
+
+			if response.Error == nil && test.expectError {
+				t.Fatal("expected error, got no error")
+			}
+
+			if response.Error != nil && !test.expectError {
+				t.Fatalf("got unexpected error: %s", response.Error)
 			}
 		})
 	}