+++ stringvalidator/one_of_example_test.go
@@ -6,6 +6,7 @@ package stringvalidator_test
 import (
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/function"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 )
 
@@ -16,6 +17,20 @@ func ExampleOneOf() {
 			"example_attr": schema.StringAttribute{
 				Required: true,
 				Validators: []validator.String{
+					// Validate string value must be "one", "two", or "three"
+					stringvalidator.OneOf([]string{"one", "two", "three"}...),
+				},
+			},
+		},
+	}
+}
+
+func ExampleOneOf_function() {
+	_ = function.Definition{
+		Parameters: []function.Parameter{
+			function.StringParameter{
+				Name: "example_param",
+				Validators: []function.StringParameterValidator{
 					// Validate string value must be "one", "two", or "three"
 					stringvalidator.OneOf([]string{"one", "two", "three"}...),
 				},